+++ app/src/main/java/com/keronei/weatherapp/data/model/Hourly.kt
@@ -1,16 +1,16 @@
 package com.keronei.weatherapp.data.model
 
 data class Hourly(
-    val clouds: Int,
+    val clouds: Double,
     val dew_point: Double,
     val dt: Int,
     val feels_like: Double,
-    val humidity: Int,
-    val pop: Int,
-    val pressure: Int,
+    val humidity: Double,
+    val pop: Double,
+    val pressure: Double,
     val temp: Double,
     val uvi: Double,
-    val visibility: Int,
+    val visibility: Double,
     val weather: List<Weather>,
     val wind_deg: Int,
     val wind_gust: Double,
